requirements.txt
absl-py==1.4.0
asgiref==3.6.0
astunparse==1.6.3
CacheControl==0.12.11
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
Deprecated==1.2.13
Django==4.2.1
firebase-admin==6.1.0
flatbuffers==23.3.3
fonttools==4.39.3
gast==0.4.0
gcloud==0.18.3
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.0
grpcio-status==1.54.0
h5py==3.8.0
httplib2==0.22.0
idna==2.8
jax==0.4.8
joblib==1.2.0
jwcrypto==1.4.2
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
ml-dtypes==0.1.0
msgpack==1.0.5
numpy==1.23.5
oauth2client==4.1.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
plotly==5.14.1
proto-plus==1.22.2
protobuf==4.22.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.17
PyJWT==2.6.0
pyparsing==3.0.9
Pyrebase4==4.6.0
python-dateutil==2.8.2
python-jwt==4.0.0
pytz==2023.3
requests==2.22.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.9
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sqlparse==0.4.4
tenacity==8.2.2
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.25.11
Werkzeug==2.3.3
wrapt==1.14.1
